requirements.txt
gym==0.26.2
stable-baselines3==2.2.1
torch>=2.1.0
numpy>=1.24.3
matplotlib>=3.7.2
tensorboard>=2.14.0
pyyaml>=6.0.1 